Ensure we pass the right Accept header for verify requests
The specification says that we need to pass application/vnd.git-lfs+json
in the Accept header for verify requests. However, we were telling the
client to pass an invalid Accept header with a different MIME type.
Since the client should know to pass the proper MIME type, and we don't
check the Accept header for verify endpoints, simply avoid telling the
client to pass an Accept header at all.