README: be more verbose about requirements
Our current description of how to use and access the administrative
interface is too terse and confusing.  People are prone to start the
server for the first time without any environment variables, which leads
to an unusable configuration.

Tell users that they must create users before they can use the server
with the client.  Be explicit about what happens if the administrative
user and password variable are not set, and try to make it clearer that
the default is that they are not set, not the literal string "unset" as
a default credential.  The server does not have default credentials,
since doing so is not a secure configuration.