Drop vendoring
Vendoring is not useful in this project, since it bloats the repository
and there's no need to be able to build without access to the Internet
at large.  Users who want to do that can run `go mod vendor` to vendor
the data once downloaded and then build offline at their will; we don't
need to explicitly vendor things ourselves.