Use embed instead of rice
For embedding assets, we've traditionally used the rice module.
However, this module always tries to read data from a zip file appended
to the binary (in an `init` function, no less) based off of reading data
from the ELF, Mach-O, or PE sections, and on Go 1.19, this segfaults.

These modules don't seem to be maintained even to the extent of
accepting pull requests, so it's not likely that this will be fixed.
Instead, let's use the Go 1.16 embed functionality, which can embed
arbitrary files into the binary simply and automatically.

Replace the rice.Box code with an embed.FS and look up paths with the
full relative path.  Bump the go.mod to 1.16 to indicate that users may
need to upgrade.  Remove the imports for rice, which is no longer
needed.